migrate to .NET 6
