+++ .github/workflows/ci.yaml
@@ -15,12 +15,6 @@ jobs:
     steps:
       - name: 🛒 Checkout
         uses: actions/checkout@v2
-      - name: 🐢 Set up MSBuild
-        uses: microsoft/setup-msbuild@v1.1
-      - name: ✨ Setup .NET 5
-        uses: actions/setup-dotnet@v1
-        with:
-          dotnet-version: 5.0.x
       - name: ✨ Setup .NET 6
         uses: actions/setup-dotnet@v1
         with:
