--- src/LJP.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.0.32112.339
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LJPcalc", "LJPmath\LJPcalc.csproj", "{4CE52415-69E8-4BE4-827D-FDFEF7DEAEF6}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LJPcalc.test", "LJPtest\LJPcalc.test.csproj", "{6BBB885B-9EB0-4725-AF3A-412A48392254}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{4CE52415-69E8-4BE4-827D-FDFEF7DEAEF6}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4CE52415-69E8-4BE4-827D-FDFEF7DEAEF6}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4CE52415-69E8-4BE4-827D-FDFEF7DEAEF6}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4CE52415-69E8-4BE4-827D-FDFEF7DEAEF6}.Release|Any CPU.Build.0 = Release|Any CPU
-		{6BBB885B-9EB0-4725-AF3A-412A48392254}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{6BBB885B-9EB0-4725-AF3A-412A48392254}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{6BBB885B-9EB0-4725-AF3A-412A48392254}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{6BBB885B-9EB0-4725-AF3A-412A48392254}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {DBCAAB22-2FDA-41DF-A28C-B5A307C466BE}
-	EndGlobalSection
-EndGlobal
