+++ src/LJPcalc.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32505.426
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LJPcalc.Core", "LJPcalc.Core\LJPcalc.Core.csproj", "{E7A223FA-3D73-477B-92EC-1377255A2CAE}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "LJPcalc.Test", "LJPcalc.Test\LJPcalc.Test.csproj", "{11168132-789F-4576-96F6-BC33D1C51667}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{E7A223FA-3D73-477B-92EC-1377255A2CAE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7A223FA-3D73-477B-92EC-1377255A2CAE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7A223FA-3D73-477B-92EC-1377255A2CAE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E7A223FA-3D73-477B-92EC-1377255A2CAE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{11168132-789F-4576-96F6-BC33D1C51667}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{11168132-789F-4576-96F6-BC33D1C51667}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{11168132-789F-4576-96F6-BC33D1C51667}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{11168132-789F-4576-96F6-BC33D1C51667}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {4707629B-567D-4C3D-B897-28ED987C7499}
+	EndGlobalSection
+EndGlobal
